pyproject.toml
[project]
name = "bmecatExplorer"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "alembic>=1.14.0",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.0",
    "lxml>=6.0.2",
    "openai>=1.10.0",
    "opensearch-py>=3.1.0",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require PostgreSQL/OpenSearch)",
    "smoke: Smoke tests (require running API server)",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = [".venv", ".tox", "build", "dist", "data"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/api/routes/search.py" = ["B008"]
"frontend/app.py" = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["src", "frontend"]
